#108 - Unable to create Sublease
Issue detected on Test. The rrr_nr_seq was changed to support
allocating up to a 6 digit number, but the business rule
for rrr.nr was not updated to reflect the change. This meant
all new rrr's where assigned a number of 180813-#### due to
format overflow. Issue has been corrected on both Test and
production.

This issue affected all transactions where more than one rrr
was added regardless of the rrr type.